src/main/java/org/leetcode/solutions/MinimumNumberOfArrows.java
package org.leetcode.solutions;

import java.util.Arrays;
import java.util.Comparator;

public class MinimumNumberOfArrows {
    public int findMinArrowShots(int[][] points) {
        if (points.length == 0) return 0;
        Arrays.sort(points, Comparator.comparingInt(o -> o[1]));

        int arrows = 1;
        int xStart, xEnd, firstEnd = points[0][1];
        for (int[] p: points) {
            xStart = p[0];
            xEnd = p[1];
            if (firstEnd < xStart) {
                arrows++;
                firstEnd = xEnd;
            }
        }
        return arrows;
    }
}
